dlfairness/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
2to3==1.0
absl-py==0.11.0
astor==0.8.1
asttokens==2.0.4
astunparse==1.6.3
backports.csv==1.0.7
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cheroot==8.5.0
CherryPy==18.6.0
click==7.1.2
cryptography==3.2.1
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
dill==0.3.3
feedparser==6.0.2
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
joblib==0.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.3
matplotlib==2.1.0
more-itertools==8.6.0
mysqlclient==2.0.1
networkx==2.5
nltk==3.5
numpy==1.19.2
opt-einsum==3.3.0
pandas==1.1.4
Pattern==3.6
pdfminer.six==20201018
Pillow==8.0.1
portend==2.7.0
protobuf==3.13.0
pycocotools==2.0.0
pycparser==2.20
pyemd==0.5.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
sgmllib3k==1.0.0
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.0.1
tempora==4.0.1
tensorboard==1.15.0
tensorboardX==2.1
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
zc.lockfile==2.0
zipp==3.4.0